+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2023-12-02
+
 ### Added
 
 - Added `traverse` method to `Input`
@@ -50,7 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `Input` struct for reading puzzle input
 
-[unreleased]: https://github.com/maxdeviant/adventurous/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/maxdeviant/adventurous/compare/v0.2.0...HEAD
+[0.1.0]: https://github.com/maxdeviant/adventurous/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/maxdeviant/adventurous/compare/v0.0.3...v0.1.0
 [0.0.3]: https://github.com/maxdeviant/adventurous/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/maxdeviant/adventurous/compare/v0.0.1...v0.0.2